modify copy.py

+++ other/copy.py
@@ -52,24 +52,29 @@ def main():
     demo_name: str = find_demo(os.getcwd())# Find src name
     if not demo_name:
         return print('local.exe not found')
-    remote: str = find_demo(path['remote'])# Find dst name
-
-    # Check dst file
-    if not remote:
-        print('remote.exe not found')
-
-    if remote == demo_name:
-        os.remove(path['remote'] + '\\' + remote)
-        return print('remote.exe == local.exe')
+    try:
+        remote: str = find_demo(path['remote'])# Find dst name
+    except:
+        remote = 'not found'
 
     shutil.copy(demo_name, path['local'], follow_symlinks=True)
-    print('Remove: ' + remote)
-    os.remove(path['remote'] + '\\' + remote)
 
-    shutil.copy(demo_name, path['remote'], follow_symlinks=True)
-    print('Copy to: ' + path['remote'])
-    os.rename(path['remote'] + '\\' + demo_name, path['remote'] + '\\' + remote)
-    print('Rename: ' + demo_name + ' to ' + remote)
+
+    # Check dst file
+    if remote != 'not found':
+        print('Remove: ' + remote)
+        os.remove(path['remote'] + '\\' + remote)    
+        if not remote:
+            print('remote.exe not found')
+        if remote == demo_name:
+            os.remove(path['remote'] + '\\' + remote)
+            return print('remote.exe == local.exe')
+
+        shutil.copy(demo_name, path['remote'], follow_symlinks=True)
+        print('Copy to: ' + path['remote'])
+        os.rename(path['remote'] + '\\' + demo_name, path['remote'] + '\\' + remote)
+        print('Rename: ' + demo_name + ' to ' + remote)
+
     print('Copy success')
